FAB-5166 Docs should use Hyperledger Fabric
There are still several places in the docs
which do not properly use Hyperledger Fabric
when referring to Hyperledger Fabric.

While this change covers a lot of files, it
simply changes all reference to
Hyperledger Fabric or provides a minor rewrite
to avoid using the terms at all.

As part of this, also addressed
- FAB-5014
- FAB-5139
- Changed docker to Docker as appropriate
- Other minor cleanups since this included most of the docs

Change-Id: I7818a44b1411abb536a595c537202615bf901199
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>