[FAB-5262] Rm committer from ProcessConfigMsg
To simplify the removal of the filter.Committer from the blockcutter, it
was pushed into the ProcessConfigMsg method, so that the consenter was
still able to supply a committer to the WriteBlock method.

However, the filter.Committer concept is going away, so this must be
removed from the consensus code.  This pushes it one level closer to
WriteBlock to facilitate its ultimate removal.

Change-Id: Icff52bb2b8a9fad96d41487147400abd149ab2f1
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>