[FAB-5264] Move interfaces from multichannel
The multichannel (formerly multichain) package became a bit of a dumping
ground for all of the interface definitions used in the orderer system.
this makes the multichannel package clunky, difficult to mock, and
encourages poor separation between components.

This CR extracts the interface definitions and moves them to more
sensible locations based on the natural separation of function within
the orderer codebase.

It is also fixed the multichannel package to generally return pointers
to structs, rather than to interfaces, following the golang best
practice of accepting interfaces and returning structs.

Change-Id: Iaf004e1dadf7bf92d106bd7c90f244e0089b9924
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>