[FAB-5266] Replace Enqueue with Order/Configure
The current consenter interface only allows for one sort of message
ingress.  All messages are received via 'Enqueue', and treated
identically.

In order for the consenter to be able to differentiate and handle config
vs non-config messages differently, the consenter needs two diferent
ingress points for messages.

This CR replaces the Enqueue method with two methods: Order and
Configure.

For the time being, these methods behave exactly as Enqueue, but will be
leveraged in future CRs.

Change-Id: I3701e5e3c0de4833a455c49acebbad70c6ed763c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>