[FAB-5271] Solo to new orderer validation flow
With all of the preliminary work finally done.  It is safe for the solo
consensus code to skip the second round of validation on transactions,
so long as the configuration sequence for the channel has not advanced.

This CR checks the current configSeq when processing a message, and if
it has not advanced, commits it directly.  If the confiSeq has advanced,
only then does it revalidate the message, dropping it if it is invalid,
or simply committing it otherwise.

Very unscientific tests on my local laptop saw throughput on Broadcast
for solo jump from 900 tps to 2000 tps.

Change-Id: Ic7340cdb66631539df3826c7a52844a53849b3ad
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>