[FAB-5406] Mutual TLS in chaincode service-P1
This commit adds the needed infrastructure to implement an interceptor
that would wrap the existing chaincode service (chaincode Support struct)
and would check that the chaincode name in the REGISTER message matches
the mapping that was set at the chaincode launch.

The full details of the flow is in the JIRA item.

Change-Id: I2a8d33fc33adf845984b73e3ab1010c34914c716
Signed-off-by: yacovm <yacovm@il.ibm.com>