[FAB-5550] configtxgen use protolator JSON parsing
For v1, configtxgen used some hacks into the configtx core code to get a
JSON representation of the config.  This CR switches the JSON encoding
to use the same protolator helpers as the configtxlator code.

The existing method was (as already noted) hacky, and not as expressive
as the configtxlator method (which also expands other opaque fields like
MSP principals and MSP definitions.  Plus, it is unnecessary code to
maintain.

Change-Id: Ibfe79de834bae6d91e7d67a9f528564658650902
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>