[FAB-5552] Fix some TODOs in msgprocessor
The restructuring of the msgprocessor code has left some outstanding
TODOs.  In particular, for conditions where a CONFIG_UPDATE is rewrapped
into a containing message, like a CONFIG or an ORDERER_TRANSACTION, this
wrapping message could end up violating rules for the channel, like the
maximum message size.  This could see the message silently dropped by
the consenter (silent from a client perspective).

This CR adds a second pass of filtering in the message flow which
generates a wrapping message.

Also fixes a small comment bug from a previous CR.

Change-Id: Ic482a8195720992c68e9986a1ff7f34004d5986c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>