[FAB-5654] SideDB - Tx simulation/validation/commit
This CR modifies the tranaction simulation, validation, and commit
code and delivers the end-to-end transaction flow that treats the
private data in a special manner. This CR mainly leverages the earlier
submitted independent CRs for sidedb feature for accomplishing this behavior.

This CR also allows ledger to receive the blocks and the pvt data from
another peer on the same channel (i.e., a peer catching up via state)

This CR is exceptionally large becasue of manily two reasons

1) The way currently the code (and specially the tests) is organized in
simulation/validation/commit flow, its not easy to submit such kind
of changes independently that cuase the change in the whole transaction
processing flow.

2) This CR causes a change in the existing ledger APIs which are used widely
across other packages (specially in the tests) and hence many files are included
for fixing the broken dependencies

Change-Id: Id29575176575f4c01793efd3476b68f8364cb592
Signed-off-by: manish <manish.sethi@gmail.com>