[FAB-5590] Add proto fields for isolated_data
The channel creation transaction needs a way to specify additional data
that is not otherwise reflected in the standard config.  Still, it's
desirable to specify this information in a consistent way, so it should
be expressed using the same structures as the configuration.

This CR introduces the notion of the 'isolated_data' field to the
ConfigUpdate message.  For now, the only consumer will be the RSCC seed
data information, which will be set to the key rscc_seed_data and will
be the bytes of a common.Config message.

Because we need to be able to be able to inspect this new config message
type, we need a way to differentiate channel config from resource
config.  Thus, this CR also introduces a 'type' field to both the Config
message and the ConfigUpdate message.

Change-Id: Ida6283dbeeca466aafd9bc6d027ce46e1baee0c6
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>