[FAB-5710] Fix missing policy warnings
The policy manager currently has a hardcoded check to look for policies
of a particular name.  This was fine, when the only policy manager was
the channel policy manager, but with the resource policy manager, this
does not make sense.

Further, the log supression for the orderer channel templating was
disabled during refactorying, this enables it once more.

Change-Id: I6994425c3075c06503203132d857b5d54b9dba8c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>