[FAB-5844] handle panic on closed channel (shim side)
This is the shim side handling for the SCC panic handling
during shutdown. The fabric side was fixed in
https://gerrit.hyperledger.org/r/#/c/12393/.

Send may happen on a closed channel when the system is
shutting down. Just catch the exception and return error.

This will be a relief to CI where this appear to be happening
more frequently of late.

Change-Id: Ie2c4032faa6a21c1e07a07e583e6f37f4a31e019
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>