[FAB-5809] Add bundlesource to track current bndl
Since the channelconfig bundle is immutable, but the channel config can
change, it's necessary to have an encapsulating structure which retains
a reference to the current immutable bundle.  Because the configuration
may be referenced by multiple go routines, it's important that access to
the reference be synchronized, in this case, through the sync/atomic
Load/Store operations.

This CR adds a BundleSoure which retains a reference to a Bundle.  It
implements the standard interface for channel configuration, but always
references the config through the atomic pointer.  A caller may also
request a static version of the bundle, if multiple coherent config
calls must be made (such as listing organizations, then looking up their
MSP info in the MSP manager).

Change-Id: I825ab93fd1b769248b44b85e68412513051b54f4
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>