[FAB-5810] Move orderer to new channelconfig bndl
The first four CRs in this series have been creating a paralllel
implementation of the channel configuration, which no components are
using.  The implementation should be compatible with the existing
implementation, so components can be swapped out individually until
there are no more references to the old model, and then it can be
reomoved.

This CR moves the orderer to the newer immutable channelconfig bundle.

Change-Id: I96cdebfa2996dede40ddacb5cdbb7891d1a29284
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>