[FAB-5813] Remove unused configtx.Manager function
This is a step in the series of moving the channel config from a mutable
structure, to an immutable one.  Thanks to the removal of concerns for
the configtx.Manager which need now only validate, not apply config,
there is a nice chunk of complicated dead code which can be removed.

Change-Id: Ic0f6fef509afcd7fe1fa5d2e484dbbdd2dcc9e15
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>