[FAB-5815] Make resources config immutable
Although very recently introduced, the resources config followed the
same mutable model as the channel config.  This leads to some awkward
looking code with far  more lines than are actually needed.

This CR converts the resources config to follow the same bundle model
that the channel config does.  It does not intrduce a bundle source like
the channel config because for the moment, the resource config may not
be changed.

Change-Id: I35614c5f7f6bee3b9ab2848033a1834ef8c76383
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>