[FAB-5818] Make MSPConfigHandler immutable
Although the MSPConfigHandler has not followed the configtx value
interface for some time, it still follows the transactional model of
Being/Propose/Precommit/Commit.

This is wrong for a number of reasons now:
  1. The underlying MSP Manager is immutable
  2. The MSPConfigHandler is only used by immutable config

This CR removes all of the mutable components of the code, drastically
simplifying the implementation and reducing the bug surface.

Change-Id: Ia940facbc152b2b17c944df8a2d24c080eff3013
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>