[FAB-5874] Support for queries over pvtdata
This CR allows queries over pvt data in addition to the
key based (Get/Set) operations. The queries over pvt
data is allowed in read-only transactions. Because, for
the peers that do not have pvt data, there is no easy way to
detect phantom items during commit time and hence guaranteeing
the serializability of trans.

If a tran performs both 'queries over pvt data' and
update operations an error is raised during simulation.

Change-Id: Ia607c13d16a20e939bbe978476757086866fa84a
Signed-off-by: manish <manish.sethi@gmail.com>