[FAB-5899] unable to use signed install package
The check for detecting if a chaincode lifecycle operation (install
instantiate or upgrade) was on a Java chaincode was expecting a
ChaincodeDeploymentSpec (currently only form supported by SDKs). This
prevented signed packages installed from CLI from working.

This is good to get working in anticipation of full future adoption
by SDKs.

Change-Id: Iead28f60e13e16048cd01464bd471e9c4d6e4d05
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>