[FAB-5284] Augment kafka message with new types
With new orderer message flow, we need to be able to differentiate
kafka message of type Normal and Config. Also we need to embed
ConfigSeq into both type of messages.

This patch augments `KafkaMessageRegular` to include these fields.
For compatibility with v1.0.x orderer, we keep default class `UNKNOWN`,
which would be auto-filled if the orderer is not aware of new types.

This is part of the effort to move Kafka-based OSNs to new message
flow.

Change-Id: I1efc569155e6a114d2e170affe333b84bf0b16c2
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>