[FAB-830] - enable transient field in CLI
This code change lets users of the golang CLI set the transient field in the
transaction proposal message.

While this change set is of idependent interest (because it is useful to
quickly set the transient field for debug purposes using the CLI), it is deeply
related to testing and debugging chaincode-level encryption: the transient
field is indeed one of the recommended ways to pass cryptographic key material
to chaincodes for the purpose of encrypting/decrypting state, arguments or
results.

Change-Id: I1f4f0dd9bf500cecabe525c00e867874577c762d
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>