FAB-5930 Upgrade to grpc-go v1.5.2
This updates grpc-go from version
1.4.0 to the latest version which is
1.5.2.  There are a few things we can
take advantage of in the future but it's
always a good idea to say current.

Most of the CR is just updating the
vendoring of the grpc and grpclog
packages.

A couple of changes where required to
some test files because v1.5.2 does not
allow you to register a service on a
running gRPC server.

Change-Id: Ia1499897426db9383801a5e717746f5c58b362e0
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>