[FAB-5753] Deduplicate identities in cauthdsl
The cauthdsl policy evaluation by spec deduplicates the identities prior
to evaluating the policy.  However, the deduplication portion of the
spec was never implemented.

Some hacky deduplication was added in the VSCC path for some reason, but
this is the right place to fix it.

Although this CR does mean that it is possible the results of policy
evaluation between versions will be different, the actual exposure to
non-determinism is quite low, as the only typical place where multiple
signatures are allowed is in endorsement which is already covered by the
VSCC hack.

There is some minor exposure because the config processing allows
multiple signatures, however, none of the default policies repeat the
same principal, so unless the config has been customized with odd
policies, this should also be a non-issue.

Change-Id: I5b25a7d4484b5e5f585245b9d4f652a0ab1730b9
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>