[FAB-4980] Account for the pvtdata during recovery
This CR modifies the recovery code for using the pvt data from the
pvtdata storage to update pvtdata state alongwith the pubdata state

The main change is in the existing test. The test now uses the pvt data also.
In addition, now the test code always runs for the historydb and
is restructured to reduce the verbosity

Change-Id: I01bc04f044230d96a5d4ab77d82ed8b70f99fa42
Signed-off-by: manish <manish.sethi@gmail.com>