[FAB-5284] Refactor methods to be member function
We need `processRegular` to be member function of class `chainImpl` so
it could access other member function more easily without passing class
pointer into the method. This is to prepare for following patches,
where re-processed messages need to be re-submitted via `Order` or
`Configure` method. Also, we don't need to pass the pointer to other
metadata in `chainImpl` anymore, i.e. lastCutBlockNumber.
`processConnect` and `processTimeToCut` are changed for consistency.

Change-Id: If386eed850fc207e0a5263ee56e36408cd4c0750
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>