[FAB-6018] Make handshake sign even when no TLS
The private data pull needs to be able to determine if a peer
is eligible of receiving a private RW set of a certain collection
by evaluating a policy.
The policy needs a signature and some signed message.

If TLS is used, the peer can use the Authentication info from the handshake,
but if the peer doesn't use TLS - the authentication info isn't populated.

This commit makes the handshake sign anyway the connection established
message, and removes the "IsAuthenticated" method.

Note: This doesn't mean, of course - that not using TLS is secure now.
It only means that the handshake message is signed, but that doesn't
imply that an attacker cannot do a MITM if TLS isn't used.

Change-Id: I376abda11060c1fe69c2b0b64d6b3d115f0114aa
Signed-off-by: yacovm <yacovm@il.ibm.com>