[FAB-6070] Add orderer capabilities structures
Certain capabilities may be required of the orderer which are not
required by the rest of the system.  For instance, changing the way a
channel creation request is validated would have an affect on ordering,
but should not have an effect on processing by the peer.  This CR adds
structures to support this.

Change-Id: I4a94a1b8f657f0b56296801672a02717d2162f9a
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>