[FAB-6071] Add application capabilities structures
Certain capabilities may be required of the peer which are completely
irrelevant to the other pieces of the system (such as the orderer).  For
instance, whether a peer may process side-DB transactions had no impact
on the orderer, because the orderer is left out of these transactions by
design.

This CR adds basic support for application/peer capabilities though none
are currently implemented.

Change-Id: I27fe1266b370b6d71791c5782bbe641c081e8f62
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>