[FAB-6087] Add configtxgen capabilties support
As a prerequisit to any real usage of the new capabilities support, it
must be possible for an admin to generate channel creation requests and
or bootstrap the ordering system network with the new configuration
keys.

This CR adds support to configtxgen to do this, though there are no
capabilities actually supported yet.

Change-Id: I9ab962727e446f593ccd262cb31ac89357cdf9d9
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>