[FAB-6137] Add simple config value util functions
In the earlier days of configuration, it was not entirely clear what
configuration would look like, so the notion of configuration templates
was invented.  This worked well to abstract the underlying
implementation away from the consumers, but now that the configuration
form has settled down, it provides an unnecessary complication that
serves to obfuscate and make understanding of the configuration more
challenging.

This CR adds a util function for directly generating the proto material
necessary to construct a *cb.ConfigValue.  There are currently no
consumers, but new ones will be added later in the change series.

Change-Id: I3a52b85a01e0387f868bb6b51f8743092655c676
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>