[FAB-6142] Move MakeChainCreationTranscation
Presently, MakeChainCreationTransaction is a function of the
channelconfig package.  Although it does seem related to the
channelconfig and references many of the constants defined there, it is
actually unneeded during the runtime for either the orderer or peer
binaries, and is only used by CLI and tests.

Therefore, it is a better fit in the configtxgen/encoder package.

This CR also renames this method to MakeChannelCreationTransaction and
switches it to use the newer underlying config group generation
introduced earlier in this series.

Change-Id: Iece267073dd612c02cc8d1b2d7feea3c02731d8f
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>