[FAB-6143] Cleanup dead code
Now that almost all references the config templating code is gone, it
should be removed entirely from the codebase.

This final CR in the channel config generation cleanup series shows a
net decrease of over 400 lines, with an even larger shift of lines of
code from the orderer/peer production binaries and into the configtxgen
binary.

Change-Id: I614a10113c0f90b822da5103319a91f2f28b66b5
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>