[FAB-6103] Add printOrg to configtxgen
One of the most common questions when attempting to add an organization
to a channel is how to get the JSON definition for that organization.
The only real way to do this at present is to generate a genesis block,
decode the block, then go look for the org definition.

This CR adds a '-printOrg' flag to configtxgen which will print the
organization's definition as JSON to make this process easier.

Change-Id: Id12098f36feae8ca4751b82afaf2914166077e50
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>