[FAB-830] Sample cc-level deterministic encryption
This change set shows how BCCSP/encshim can be used to encrypt data in a way
where the IV is also specified through the transient field. This is useful
when a chaincode that uses encshim needs to be endorsed by multiple peers.
Without this change-set, the KVS writes would differ across the endorsing
peers and so endorsement would fail. With this change set, the invoker can
specify the IV and so encryption can produce repeatable KVS writes that can
be endorsed by multiple peers.

Change-Id: I0f081926d2a57d553f91c9816669ddbc76a77208
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>