[FAB-6426] Remove Capability msg 'required' field
The Capability msg currently only has a single field: 'required'.  This
field doesn't make a ton of sense, because if a capability were not
required, then it could simply not be added to the capabilities map.  In
the interest of not allowing the expression of the same idea two ways
(and the complexity which comes with it), this CR removes the 'required'
field from the Capability message and instead makes it an empty message
which may be extended in the future, should the need ever arise.

Change-Id: I4c48d2a3d0c4fedb4bcf86f8476f34470ad494c5
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>