[FAB-6449] Add upgrade fix for empty mod_policy
Thanks to bugs

  FAB-5309
  FAB-6080

when channels were constructed with a v1.0.x orderer, the mod_policy for
certain elements was incorrectly set as empty.  This has the effect that
no changes to those elements may ever be made.  Because these elements
must be able to be modified in order to complete a channel upgrade to
v1.1, there needs to be an upgrade path to fix chains which were
constructed in the broken manner as described by the above bugs.

This CR adds an admittedly very hacky fix in the config processing code.
It in effect creates a one time special case migration, that when
Capabilities are added at the channel orderer level, any empty
mod_policy is recomputed to be set to "Admins".

This is a hack because it violates layering, introduces knowledge of config
structure into the configtx manager, and, would never have been designed
this way.  However, it is also a minimally invasive change which is easily
tested and isolated, so is hopefully an acceptable solution.

Change-Id: Ic27326e50d7ff27a1c6e2c796404d546653258e5
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>