[FAB-6462] Enforce PEM types for gossip identities
The MSP support for gossip (MessageCryptoService) takes into account
The entire PEM of an x509 based identity, which is too lenient.

Instead of enforcing explicit sanitation of input for GetPKIidOfCert(),
We will just make identities that have the wrong PEM type - invalid.

This is much more efficient because the GetPKIidOfCert is called many times
in the code, and sanitizing its input is therefore costly,
while if we just make identities that have the wrong type - invalid,
it would implicitly force sanitation of identities.
Another argument for doing so, would be that the Validate() is costly
anyway and is rarely performed in the peer's life cycle, therefore -
adding an additional slight overhead to it doesn't affect much the cost
of validating an identity.

The BCCSP-MSP Serialize() method already puts the right PEM type as of
an earlier commit that is targeted for v1.1 and isn't in v1.0.x

This change set makes the MessageCryptoService consider x509 based
identities valid only if they have a PEM type of either CERTIFICATE
or a blank one.

Change-Id: Ic332f592cf2ebfe364f6c99ae762b2789d9c2b51
Signed-off-by: yacovm <yacovm@il.ibm.com>