[FAB-6620] Prevent private data send in instantiate
Currently the collection configurations are stored in the LSCC namespace.
Therefore, if an instantiate transaction contains private data -
it's impossible to obtain the collection configuration in order to
know to which peers to disseminate to, etc.

Therefore we should just disable private data dissemination upon
endorsement altogether.

This change set makes the endorsement fail if private data is
detected in the simulation results in instantiate.

Change-Id: Icb6d0e9df2505e93f2350063d43913b71e3f2761
Signed-off-by: yacovm <yacovm@il.ibm.com>