[FAB-6367] create genesis tx/block with other MSPs
This change set introduces an 'MSPType' field to configtx.yaml, so that the
type of MSP can be specified. This permits credentials for other MSP types,
most notably the idemix MSP, to be included in genesis tx/blocks so that they
can be used as well. If no MSPType is specified, the standard one is
automatically assumed.

Change-Id: I5a609b57cfe85ae030f7a8fbecf8cfd71d3074ae
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>